+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED] - 2022-##-##
+### Added
+- schedule view
+- EpisodeSchedule, EpisodeCard, and Alert
+- episode schedule for all rated/watchlisted via Drawer
+- episode schedule per show via button in ShowDetailsModal
+- streamable icon on ShowCard
+
+### Changed
+- ShowImage extracted from ShowCard
+- scroll to top on view change instead of after fetch
+- useOnScreen also provides position
+- readme usage, features, and styling
 
 ## [1.16.0] - 2022-04-17
 ### Added